Move Wrap out of Form and into Encode.
